VW: remove legacy WeConnect
