--- templates/definition/vehicle/vw-id.yaml
@@ -1,21 +1,0 @@
-template: id
-products:
-  - brand: Volkswagen
-    description:
-      generic: We Connect ID
-requirements:
-  description:
-    de: ID.3, ID.4 (funktioniert auch für e-Golf, e-Up)
-    en: ID.3, ID.4 (also works for e-Golf, e-Up)
-params:
-  - preset: vehicle-base
-  - preset: vehicle-identify
-  - name: vin
-    example: WVWZZZ...
-  - name: timeout
-    default: 10s
-render: |
-  type: id
-  {{ include "vehicle-base" . }}
-  {{ include "vehicle-identify" . }}
-  timeout: {{ .timeout }}
